feat(kilo-vscode): refactoring and organization
