+++ packages/kilo-vscode/src/legacy-migration/sessions/lib/ids.ts
