+++ packages/kilo-vscode/src/legacy-migration/sessions/lib/legacy-types.ts
