+++ packages/kilo-vscode/src/legacy-migration/sessions/lib/parts/merge-tools.ts
@@ -1,5 +1,5 @@
 import type { KilocodeSessionImportPartData as Part } from "@kilocode/sdk/v2"
-import type { LegacyApiMessage } from "../legacy-session-types"
+import type { LegacyApiMessage } from "../legacy-types"
 import { getText, getToolUse, isToolResult, record } from "./parts-util"
 
 type Body = NonNullable<Part["body"]>
