+++ packages/kilo-vscode/src/legacy-migration/sessions/lib/parts/parts-builder.ts
