+++ packages/kilo-vscode/src/legacy-migration/sessions/lib/parts/parts-util.ts
@@ -1,4 +1,4 @@
-import type { LegacyApiMessage } from "../legacy-session-types"
+import type { LegacyApiMessage } from "../legacy-types"
 
 type ToolUse = {
   type?: string
