+++ packages/kilo-vscode/src/legacy-migration/sessions/lib/session.ts
@@ -1,19 +1,24 @@
 import type { KilocodeSessionImportSessionData as Session } from "@kilocode/sdk/v2"
-import type { LegacyHistoryItem } from "./legacy-session-types"
+import type { LegacyHistoryItem } from "./legacy-types"
 import { createSessionID } from "./ids"
 
 export function createSession(id: string, item: LegacyHistoryItem | undefined, projectID: string): NonNullable<Session["body"]> {
   const session = makeSession()
 
   session.id = createSessionID(id)
+
   session.projectID = projectID
+
   session.slug = id
 
   session.directory = item?.workspace ?? ""
+
   session.title = item?.task ?? id
+
   session.version = "v2"
 
   session.timeCreated = item?.ts ?? 0
+
   session.timeUpdated = item?.ts ?? 0
 
   return session
