+++ packages/kilo-vscode/src/legacy-migration/sessions/parser.ts
@@ -1,14 +1,14 @@
-import type { LegacyHistoryItem } from "./legacy-session-types"
+import type { LegacyHistoryItem } from "./lib/legacy-types"
 import type {
   KilocodeSessionImportMessageData as Message,
   KilocodeSessionImportPartData as Part,
   KilocodeSessionImportProjectData as Project,
   KilocodeSessionImportSessionData as Session,
 } from "@kilocode/sdk/v2"
-import { createMessages } from "./create-messages"
-import { createParts } from "./parts/parts"
-import { createProject } from "./create-project"
-import { createSession } from "./create-session"
+import { createMessages } from "./lib/messages"
+import { createParts } from "./lib/parts/parts"
+import { createProject } from "./lib/project"
+import { createSession } from "./lib/session"
 
 export interface NormalizedSession {
   project: NonNullable<Project["body"]>
@@ -17,7 +17,7 @@ export interface NormalizedSession {
   parts: Array<NonNullable<Part["body"]>>
 }
 
-export async function normalizeSession(id: string, dir: string, item?: LegacyHistoryItem): Promise<NormalizedSession> {
+export async function parseSession(id: string, dir: string, item?: LegacyHistoryItem): Promise<NormalizedSession> {
   const project = createProject(item)
   const session = createSession(id, item, project.id)
   const messages = await createMessages(id, dir, item)
